[refact] lib/common.py Captcha 관련 함수 분리
